backend/fastapi/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
httpx
python-dotenv
dedalus-labs
